django-backend/requirements.txt
Django==5.1.1
djangorestframework==3.15.2
strawberry-graphql==0.242.0
strawberry-graphql-django==0.30.0
django-filter==24.3
psycopg2-binary==2.9.9
celery==5.4.0
redis==5.2.0
django-celery-beat==2.7.0
django-cors-headers==4.4.0
python-dotenv==1.0.1
stellar-sdk==12.3.0
gunicorn==23.0.0
requests==2.32.3
pytest==8.3.3
pytest-django==4.9.0
black==24.10.0
ruff==0.7.0
dj-database-url==2.2.0
